refactor(cast): share provider, hardfork and json helpers
