+++ crates/cast/src/cmd/vaddr/resolve.rs
@@ -1,8 +1,7 @@
-use crate::tempo::tempo_provider;
+use crate::{cmd::print_json_or, tempo::tempo_provider};
 use alloy_primitives::Address;
 use eyre::Result;
 use foundry_cli::opts::RpcOpts;
-use foundry_common::shell;
 use serde_json::json;
 use tempo_alloy::contracts::precompiles::{ADDRESS_REGISTRY_ADDRESS, IAddressRegistry};
 
@@ -19,20 +18,19 @@ pub(super) async fn run(addr: Address, rpc: RpcOpts) -> Result<()> {
     }
 
     let master_address = (!master.is_zero()).then(|| format!("{master}"));
-    if shell::is_json() {
-        let payload = json!({
-            "address": format!("{addr}"),
-            "master_id": format!("{}", decoded.masterId),
-            "user_tag": format!("{}", decoded.userTag),
-            "master_address": master_address,
-        });
-        sh_println!("{}", serde_json::to_string_pretty(&payload)?)
-    } else {
-        sh_println!(
+    let payload = json!({
+        "address": format!("{addr}"),
+        "master_id": format!("{}", decoded.masterId),
+        "user_tag": format!("{}", decoded.userTag),
+        "master_address": master_address,
+    });
+    print_json_or(
+        payload,
+        format!(
             "Virtual address: {addr}\nMaster ID:       {}\nUser tag:        {}\nMaster address:  {}",
             decoded.masterId,
             decoded.userTag,
             master_address.as_deref().unwrap_or("(unregistered)"),
-        )
-    }
+        ),
+    )
 }
