+++ crates/cast/src/tempo.rs
@@ -9,29 +9,34 @@ use alloy_transport::{BoxTransport, TransportConnect, TransportError};
 use eyre::Result;
 use foundry_cli::{
     json::print_json_success,
-    opts::{EthereumOpts, RpcOpts, TempoOpts},
+    opts::{EthereumOpts, TempoOpts},
     utils::{LoadConfig, get_chain},
 };
 use foundry_common::{
     FoundryTransactionBuilder,
-    provider::{ProviderBuilder, RetryProvider},
+    provider::{ProviderBuilder, RetryProvider, is_rpc_method_not_found},
     shell,
     tempo::{maybe_print_fee_token, resolve_and_set_fee_token},
 };
 use foundry_config::{Chain, Config, Eip1559FeeEstimatePreset};
+use foundry_evm::hardfork::TempoHardfork;
 use foundry_wallets::{TempoAccountsWallet, WalletOpts, WalletSigner};
+use serde::Deserialize;
 use serde_json::Value;
 use std::str::FromStr;
 use tempo_alloy::{
     TempoNetwork,
+    provider::TempoProviderExt,
     transport::{RelayConnector, SponsorshipMode},
 };
 
 pub use foundry_common::tempo::TempoSponsor;
 
-/// Loads the config for `rpc` and builds a Tempo provider from it.
-pub(crate) fn tempo_provider(rpc: &RpcOpts) -> Result<(Config, RetryProvider<TempoNetwork>)> {
-    let config = rpc.load_config()?;
+/// Loads the config for `opts` and builds a Tempo provider from it.
+pub(crate) fn tempo_provider(
+    opts: &impl LoadConfig,
+) -> Result<(Config, RetryProvider<TempoNetwork>)> {
+    let config = opts.load_config()?;
     let provider = ProviderBuilder::<TempoNetwork>::from_config(&config)?.build()?;
     Ok((config, provider))
 }
@@ -161,6 +166,75 @@ pub(crate) fn ensure_session_not_browser(tempo: &TempoOpts, browser: bool) -> Re
     Ok(())
 }
 
+/// Fails with `message` unless `hardfork` is active on the RPC.
+pub(crate) async fn require_hardfork<P: Provider<TempoNetwork>>(
+    provider: &P,
+    hardfork: TempoHardfork,
+    message: &str,
+) -> Result<()> {
+    if !is_tempo_hardfork_active(provider, hardfork).await? {
+        eyre::bail!("{message}");
+    }
+    Ok(())
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct AnvilNodeInfo {
+    hard_fork: Option<String>,
+    network: Option<String>,
+}
+
+pub(crate) async fn is_tempo_hardfork_active<P: Provider<TempoNetwork>>(
+    provider: &P,
+    hardfork: TempoHardfork,
+) -> Result<bool> {
+    match provider.is_hardfork_active(hardfork).await {
+        Ok(active) => Ok(active),
+        Err(err) if is_rpc_method_not_found(&err) => {
+            match anvil_tempo_hardfork_active(provider, hardfork).await {
+                Ok(Some(active)) => Ok(active),
+                _ => Err(err.into()),
+            }
+        }
+        Err(err) => Err(err.into()),
+    }
+}
+
+/// Fails early with `requirement` when a Tempo precompile is not active yet: a pre-fork call
+/// would succeed as a silent no-op instead of reverting. Prefers the hardfork query and falls
+/// back to checking the precompile's code when the RPC lacks the method.
+pub(crate) async fn ensure_tempo_precompile_active<P: Provider<TempoNetwork>>(
+    provider: &P,
+    hardfork: TempoHardfork,
+    precompile: Address,
+    requirement: &str,
+) -> Result<()> {
+    let active = match is_tempo_hardfork_active(provider, hardfork).await {
+        Ok(active) => active,
+        Err(_) => !provider.get_code_at(precompile).await?.is_empty(),
+    };
+    eyre::ensure!(active, "{requirement}");
+    Ok(())
+}
+
+async fn anvil_tempo_hardfork_active<P: Provider<TempoNetwork>>(
+    provider: &P,
+    hardfork: TempoHardfork,
+) -> Result<Option<bool>, TransportError> {
+    let info = provider.raw_request::<_, AnvilNodeInfo>("anvil_nodeInfo".into(), ()).await?;
+    Ok(active_from_anvil_node_info(&info, hardfork))
+}
+
+fn active_from_anvil_node_info(info: &AnvilNodeInfo, hardfork: TempoHardfork) -> Option<bool> {
+    (info.network.as_deref() == Some("tempo")).then(|| {
+        info.hard_fork
+            .as_deref()
+            .and_then(|active_hardfork| active_hardfork.parse::<TempoHardfork>().ok())
+            .is_some_and(|active_hardfork| active_hardfork >= hardfork)
+    })
+}
+
 /// Connector for reusing an already-configured RPC transport.
 ///
 /// This preserves Foundry transport behavior such as MPP payment handling when a sponsor relay is
@@ -265,6 +339,23 @@ mod tests {
     use super::*;
     use alloy_provider::{ProviderBuilder as AlloyProviderBuilder, mock::Asserter};
     use alloy_rpc_client::RpcClient;
+
+    #[test]
+    fn active_from_anvil_node_info_requires_tempo_network() {
+        let info = |network: &str, hard_fork: &str| AnvilNodeInfo {
+            network: Some(network.to_string()),
+            hard_fork: Some(hard_fork.to_string()),
+        };
+        let tempo_t3 = info("tempo", "T3");
+        assert_eq!(active_from_anvil_node_info(&tempo_t3, TempoHardfork::T2), Some(true));
+        assert_eq!(active_from_anvil_node_info(&tempo_t3, TempoHardfork::T3), Some(true));
+        assert_eq!(active_from_anvil_node_info(&tempo_t3, TempoHardfork::T4), Some(false));
+        assert_eq!(
+            active_from_anvil_node_info(&info("tempo", "T11"), TempoHardfork::T11),
+            Some(true)
+        );
+        assert_eq!(active_from_anvil_node_info(&info("ethereum", "T3"), TempoHardfork::T3), None);
+    }
 
     #[tokio::test]
     async fn sponsor_relay_reuses_existing_default_transport() {
